requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohappyeyeballs==2.3.4
aiohttp==3.9.5
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.7
django-appconf==1.0.6
django-ckeditor==6.7.1
django-etc==1.4.0
django-hitcount==1.3.5
django-image-cropping==1.7
django-jazzmin==3.0.0
django-js-asset==2.2.0
environs==11.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
marshmallow==3.21.3
MouseInfo==0.1.3
multidict==6.0.5
packaging==24.1
phonenumbers==8.13.42
pillow==10.4.0
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
python-socks==2.5.0
pytweening==1.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
telebot==0.0.5
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
